empty content file case revision needed

+++ bonus_files/exec_cmd_bonus.c
@@ -6,7 +6,7 @@
 /*   By: donglee2 <donglee2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/06 12:01:08 by donglee2          #+#    #+#             */
-/*   Updated: 2023/07/17 15:11:16 by donglee2         ###   ########seoul.kr  */
+/*   Updated: 2023/07/17 20:30:21 by donglee2         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,8 @@ static void	exec_1st_cmd(int fds[2], t_args *args, char **envp)
 	}
 	dup2(fd, STDIN_FILENO);
 	close (fd);
-	unlink(args->infile_name);
+	if (args->bonus_flag)
+		unlink(args->infile_name);
 	close(fds[0]);
 	dup2(fds[1], STDOUT_FILENO);
 	close(fds[1]);
